refactor(InitializerForm): adjust the location of the form according to its intended operation
