+++ src/pages/privileges/outlets/users/forms/GeneralInfoForm/GeneralInformationForm.stories.tsx
@@ -2,7 +2,7 @@ import { BrowserRouter } from "react-router-dom";
 import { GeneralInformationForm, GeneralInformationFormProps } from "./index";
 import { action } from "@storybook/addon-actions";
 import { userEntriesDataMock } from "@mocks/apps/privileges/users/users.mock";
-import { IGeneralInformationEntry } from "../../../types/forms.types";
+import { IGeneralInformationEntry } from "../../types/forms.types";
 import { StoryFn } from "@storybook/react";
 
 const story = {
